Introduce a PhonePlatformProvider Component
It introduces a Provider component, and an inline
editing replacement as well

It extracts the logic from SharingTooltip for editor
and preview, so SharingTooltip won't have to be duplicate

It also introduces a usePhonePlatform hook, which is
just a wrapper for useContext(PhonePlatformContext)
It allows to use usePhonePlatform freely anywhere
