+++ entry_types/scrolled/package/src/frontend/inlineEditing/components.js
@@ -7,4 +7,4 @@ export {LayoutWithPlaceholder} from './LayoutWithPlaceholder';
 export {EditableText} from './EditableText';
 export {EditableInlineText} from './EditableInlineText';
 
-export {SharingTooltip} from './SharingTooltip';
+export {PhonePlatformProvider} from './PhonePlatformProvider';
