+++ entry_types/scrolled/package/src/frontend/navigation/AppHeader.js
@@ -12,8 +12,7 @@ import {Logo} from './Logo';
 
 import styles from './AppHeader.module.css';
 
-import PhonePlatformContext from './PhonePlatformContext';
-import {browser} from 'pageflow/frontend';
+import {PhonePlatformProvider} from '../PhonePlatformProvider';
 
 export function AppHeader(props) {
   const [navExpanded, setNavExpanded] = useState(true);
@@ -24,11 +23,6 @@ export function AppHeader(props) {
 
   const ref = useRef();
   useNativeScrollPrevention(ref);
-
-  var isPhonePlatform = false;
-  if (browser.has('phone platform')) {
-    isPhonePlatform = true;
-  }
 
   const chapters = entryStructure.map((chapter) => {
     return ({
@@ -108,9 +102,9 @@ export function AppHeader(props) {
         </nav>
 
         <div className={classNames(styles.contextIcons)}>
-          <PhonePlatformContext.Provider value={isPhonePlatform} >
+          <PhonePlatformProvider>
             <SharingMenu />
-          </PhonePlatformContext.Provider>
+          </PhonePlatformProvider>
           <LegalInfoMenu />
         </div>
       </div>
