+++ entry_types/scrolled/package/src/frontend/usePhonePlatform.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import PhonePlatformContext from './PhonePlatformContext';
+
+export function usePhonePlatform() {
+    return React.useContext(PhonePlatformContext);
+}
